+++ common/utils/resolvers/index.ts
@@ -1,0 +1,13 @@
+import clientPreference from './clientPreference'
+import commentDraft from './commentDraft'
+import gatewayUrls from './gatewayUrls'
+
+export default {
+  Query: {
+    clientPreference,
+    commentDraft
+  },
+  Official: {
+    gatewayUrls
+  }
+}