+++ components/Language/LanguageContext.test.tsx
@@ -1,6 +1,6 @@
+import { render } from '@testing-library/react'
 import React from 'react'
 import { MockedProvider } from 'react-apollo/test-utils'
-import { render } from 'react-testing-library'
 
 import { LanguageConsumer, LanguageProvider } from './LanguageContext'